+++ src/kernels/static_threads.rs
@@ -25,7 +25,7 @@ impl GemmKernel for StaticIkjGemm {
         "static-ikj"
     }
 
-    fn compute(&self, lhs: &Matrix<f64>, rhs: &Matrix<f64>, output: &mut Matrix<f64>) {
+    fn compute(&self, lhs: &Matrix<f32>, rhs: &Matrix<f32>, output: &mut Matrix<f32>) {
         assert_gemm_dimensions(lhs, rhs, output);
         let n = lhs.cols();
         output.as_mut_slice().fill(0.0);
